Close mutation gaps in the Biome, copy-check and cleanup scripts
Lock the Biome package version to the one biome.json is written against,
name the copy catalog folder once and check it really holds the site's
English copy, and assert the message on the combined-failure error. Split
the copy-check tests so each one sits beside the file it tests.
